# openapi.yaml
openapi: 3.0.1
info:
  title: PRIN135-Demo-10-OpenAPI
  version: PRIN135-Demo-10-OpenAPI
  description: HEHEHEHEHEHEHEHEHEHE
  contact:
    email: gervacio2130732@mls.ceu.edu.ph
    name: Mark Gervacio
externalDocs:
  url: swagger.json
  description: swagger.json
servers:
  - url: https://demo-09-db-integration-main2.vercel.app
    variables: {}
    description: Production Link
paths:
  /tasks:
    get:
      summary: Returns all tasks.
      description: Returns all tasks.
      operationId: ''
      responses:
        '200':
          description: Task Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskGetResponse'
        '404':
          description: Task not Found
      tags:
        - Tasks
    post:
      summary: Creates a tasks.
      description: Creates a tasks.
      operationId: ''
      requestBody:
        description: 'hehehehhe'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        default:
          description: Default error sample response
      tags:
        - Tasks
  /tasks{id}:
    get:
      summary: Returns a task by specified id.
      description:  Returns a task by specified id.
      parameters:
        - name: id
          in: path
          description: 'The id of the task.'
          required: true
          schema:
            type: integer
      operationId: ''
      responses:
        '200':
          description: Task Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGetResponse'
        '404':
          description: Task not Found
      tags:
        - Tasks
    put:
      summary: Updates a task by specified id.
      description: Updates a task by specified id.
      parameters:
        - name: id
          in: path
          description: 'The id of the task.'
          required: true
          schema:
            type: integer
      operationId: ''
      requestBody:
        description: 'hahahahhahahahaha'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        default:
          description: Default error sample response
      tags:
        - Tasks
    delete:
      summary: Deletes a task by specified id.
      description: Deletes a task by specified id.
      parameters:
        - name: id
          in: path
          description: 'The id of the task.'
          required: true
          schema:
            type: integer
      operationId: ''
      responses:
        default:
          description: Default error sample response
      tags:
        - Tasks
components: 
  schemas: 
    TaskCreateRequest:
      type: object
      properties:
        name:
          type: string
    TaskUpdateRequest:
      type: object
      properties:
        name:
          type: string
        isDone:
          type: boolean
    TaskGetResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isDone:
          type: boolean    
          